+++ ball.h
@@ -2,16 +2,16 @@
 #include <gtk/gtk.h>
 
 typedef struct _BALL{
-	int x, y, r;	//x position, y position, radious
+	float x, y, r;	//x position, y position, radious
 	unsigned int width;		//line width
 	T_SPEED speed;	//move speed
 	unsigned int clr;			//color
 } T_BALL;
 
 void ball_setRadius(T_BALL *pBall, int r);
-int ball_getRadius(T_BALL *pBall);
+float ball_getRadius(T_BALL *pBall);
 
-void ball_getPos(T_BALL *pBall, int *px, int *py);
+void ball_getPos(T_BALL *pBall, float *px, float *py);
 
 T_SPEED ball_getSpeed(T_BALL *pBall);
 
@@ -30,7 +30,7 @@ void ball_move(T_BALL *pBall, int winWidth, int winHeight);
 void ball_addSpeed(T_BALL *pBall, float xSpeed, float ySpeed);
 
 //initialize a ball
-T_BALL *ball_init(int x, int y, int r, unsigned int clr);
+T_BALL *ball_init(float x, float y, float r, unsigned int clr);
 
 //destroy a ball
 void ball_destroy(T_BALL *pBall);
